Create abstract provider and use it
