+++ src/Tenancy/Support/HooksProvider.php
@@ -16,25 +16,10 @@ declare(strict_types=1);
 
 namespace Tenancy\Support;
 
-use Illuminate\Support\ServiceProvider;
-use Tenancy\Lifecycle\Contracts\ResolvesHooks;
+use Tenancy\Providers\Provides\ProvidesHooks;
+use Tenancy\Support\Provider;
 
-abstract class HooksProvider extends ServiceProvider
+abstract class HooksProvider extends Provider
 {
-    /**
-     * Lifecycle event hooks. Hooks that run specific logic
-     * during Tenant creation, updates or deletion.
-     *
-     * @var array
-     */
-    protected $hooks = [];
-
-    public function register()
-    {
-        $this->app->resolving(ResolvesHooks::class, function (ResolvesHooks $resolver) {
-            foreach ($this->hooks as $hook) {
-                $resolver->addHook($hook);
-            }
-        });
-    }
+    use ProvidesHooks;
 }
